implemented palettes lookup computation fixed index mistakes

+++ naive_defs.hpp
@@ -1,11 +1,13 @@
 #ifndef NAIVE_DEFS_HPP
 #define NAIVE_DEFS_HPP
 
+#include <cstdint>
+
 namespace constants
 {
-    const int PALETTE_ENTRIES = 256;
-    const int ALPHA_LEVELS = 16;
-    const int LIGHT_LEVELS = 256;
+    const uint16_t PALETTE_ENTRIES = 256;
+    const uint16_t ALPHA_LEVELS = 16;
+    const uint16_t LIGHT_LEVELS = 256;
 }
 
 #endif // NAIVE_DEFS_HPP
